python/requirements.txt
argparse==1.2.1
pyserial==3.2.1
requests==2.11.1
wsgiref==0.1.2
